[fix][broker] Keep Shared permit fix accounting-only
Restore the existing executor and read-trigger behavior while retaining exact Consumer identity and rejected-submission accounting coverage. Remove the threading-specific tests and keep the behavior matrix for modern/classic Shared and Key_Shared dispatchers.
